Create breakfile
This function will take a path and file for a csv file (entered in quotes, with / to divide the path beginning at one's default R path) and an integer and will break that file into smaller files of that number of records. If there are records remaining, then the final file will contain the remaining records.